Abstract Homepage component to own directory
