+++ components/Homepage/Homepage.styles.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Title = styled.h1`
+  font-family: helvetica;
+  color: darkblue;
+`;
