+++ components/Homepage/index.ts
@@ -1,0 +1,1 @@
+export { default as Homepage } from "./Homepage";
