+++ pages/index.tsx
@@ -1,40 +1,10 @@
-import React from "react";
 import fetch from "isomorphic-unfetch";
-import styled from "styled-components";
+import { Homepage } from "../components/Homepage";
+import { HomepageProps as IndexProps } from "../components/Homepage/types";
 
-const Title = styled.h1`
-  font-family: helvetica;
-  color: darkblue;
-`;
-
-type TeaserContent = {
-  id: string;
-  title: string;
-  excerpt: string;
-};
-
-type TeaserProps = {
-  content: TeaserContent;
-};
-
-type MainProps = {
-  data: TeaserContent[];
-};
-
-const Teaser = ({ content }: TeaserProps) => (
-  <article>
-    <h2>{content.title}</h2>
-    <p>{content.excerpt}</p>
-  </article>
-);
-
-const Index = ({ data }: MainProps) => (
+const Index = ({ data }: IndexProps) => (
   <div>
-    <Title>The Broken Banana</Title>
-    <p>An online magazine for troubled bananas</p>
-    {data.map(item => (
-      <Teaser key={`teaser-${item.id}`} content={item} />
-    ))}
+    <Homepage data={data} />
   </div>
 );
